backend/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
psycopg2-binary==2.9.10
redis==6.4.0
celery==5.5.3
django-celery-beat==2.7.0
flower==2.0.1
scikit-learn==1.7.1
numpy==2.3.2
pandas==2.3.1
matplotlib==3.10.5
seaborn==0.13.2
requests==2.32.3
python-decouple==3.8
django-cors-headers==4.6.0
pywebpush==1.14.1

